Crate wasmedge_sdk

Overview

WasmEdge Rust SDK provides idiomatic Rust language bindings for WasmEdge

Notice: This project is still under active development and not guaranteed to have a stable API.

Documentation

WasmEdge website

WasmEdge GitHub Page

WasmEdge Rust SDK GitHub Page

WasmEdge Rust SDK Examples

Get Started

Since this crate depends on the WasmEdge C API, it needs to be installed in your system first. Please refer to WasmEdge Installation and Uninstallation to install the WasmEdge library. The versioning table below shows the version of the WasmEdge library required by each version of the wasmedge-sdk crate.

wasmedge-sdk	WasmEdge lib	wasmedge-sys	wasmedge-types	wasmedge-macro	async-wasi

0.9.0	0.13.0	0.14.0	0.4.2	0.4.0	0.0.1

0.8.1	0.12.1	0.13.1	0.4.1	0.3.0	-

0.8.0	0.12.0	0.13.0	0.4.1	0.3.0	-

0.7.1	0.11.2	0.12.2	0.3.1	0.3.0	-

0.7.0	0.11.2	0.12	0.3.1	0.3.0	-

0.6.0	0.11.2	0.11	0.3.0	0.2.0	-

0.5.0	0.11.1	0.10	0.3.0	0.1.0	-

0.4.0	0.11.0	0.9	0.2.1	-	-

0.3.0	0.10.1	0.8	0.2	-	-

0.1.0	0.10.0	0.7	0.1	-	-

WasmEdge Rust SDK can automatically search the following paths for the WasmEdge library:

$HOME/.wasmedge (Linux/macOS)

/usr/local (Linux/macOS)

If you have installed the WasmEdge library in a different path, you can set the WASMEDGE_INCLUDE_DIR and WASMEDGE_LIB_DIR environment variables to the path of the WasmEdge library.

Notice: The minimum supported Rust version is 1.68.

Examples

The Examples of WasmEdge RustSDK repo contains a number of examples that demonstrate how to use the WasmEdge Rust SDK.

Contributing

Please read the contribution guidelines on how to contribute code.

License

This project is licensed under the terms of the Apache 2.0 license.

Modules

async

Defines the types used in the async scenarios.

config

Defines the structs used to construct configurations.

dock

Defines WasmEdge VmDock and Param types.

error

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge error types.

plugin

Defines plugin related structs.

types

Defines the general types.

wasi

Defines wasi module instance types, including WasiInstance, WasiNnInstance, wasi-crypto instances.

Macros

params

Used to pass arguments to the following methods:

Structs

Caller

Represents the calling frame on top of stack.

Compiler

Defines WasmEdge ahead-of-time(AOT) compiler and the relevant APIs.

CoreVersion

The version info of WasmEdge core

Driver

Defines WasmEdge Driver functions

Executor

Defines an execution environment for both pure WASM and compiled WASM.

ExportType

Defines the types of the exported instances.

Func

Defines a host function instance.

FuncRef

Defines a reference to a host function.

FuncType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge FuncType.

FuncTypeBuilder

Defines a type builder for creating a FuncType instance.

Global

Defines a WebAssembly global variable, which stores a single value of the given GlobalType and a flag indicating whether it is mutable or not.

GlobalType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge GlobalType.

ImportObject

Defines an import object that contains the required import data used when instantiating a module.

ImportObjectBuilder

Creates a normal or wasi import object.

ImportType

Defines the types of the imported instances.

Instance

Represents an instantiated module.

LogManager

Manipulates the runtime logger.

Memory

Defines a linear memory.

MemoryType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge MemoryType.

Module

Defines compiled in-memory representation of an input WASM binary.

Statistics

Used to collect statistics of the WasmEdge runtime, such as the count of instructions in execution.

Store

Represents all global state that can be manipulated by WebAssembly programs. A store consists of the runtime representation of all instances of functions, tables, memories, and globals.

Table

Defines a table storing the references to host functions or external objects.

TableType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge TableType.

Vm

A Vm defines a virtual environment for managing WebAssembly programs.

VmBuilder

Constructs a Vm instance.

Enums

CompilerOptimizationLevel

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge AOT compiler optimization level.

CompilerOutputFormat

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge AOT compiler output binary format.

ExternalInstanceType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines the type of external WasmEdge instances.

HostRegistration

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge host module registration enum.

Mutability

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines the mutability property of WasmEdge Global variables.

RefType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge reference types.

ValType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge value types.

Traits

AsInstance

The object used as an module instance is required to implement this trait.

Engine

The object that is used to perform a host function is required to implement this trait.

WasmVal

Defines the function converting a value of Rust type to the one of Wasm type.

WasmValType

Describes the mapping of Rust type to Wasm type.

WasmValTypeList

Describes the mapping of a tuple of Rust types to Wasm types.

Functions

wat2wasm

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module.

Type Definitions

Finalizer

Defines the signature of a finalizer funtion that is used to free the host data.

HostFn

Defines the signature of a host function.

NeverType

This is a workaround solution to the never type in Rust. It will be replaced by ! once it is stable.

WasmEdgeResult

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module The WasmEdge result type.

WasmValue

WebAssembly value type.

Attribute Macros

async_host_function

Declare a native async function that will be used to create an async host function instance.

host_function

Declare a native function that will be used to create a host function instance.