wasmedge_sdk::async::vm
Struct Vm

pub struct Vm<'inst, T: ?Sized + Send + AsyncInst> { /* private fields */ }

A Vm defines a virtual environment for managing WebAssembly programs.

Example

The example below presents how to register a module as named module in a Vm instance and run a target wasm function.

// If the version of rust used is less than v1.63, please uncomment the follow attribute.
// #![feature(explicit_generic_args_with_impl_trait)]
#[cfg(not(feature = "async"))]
use wasmedge_sdk::{params, VmBuilder, WasmVal, wat2wasm, ValType, NeverType};

#[cfg_attr(test, test)]
fn main() -> Result<(), Box<dyn std::error::Error>> {
    #[cfg(not(feature = "async"))]
    {
        // create a Vm context
        let vm = VmBuilder::new().build()?;

        // register a wasm module from the given in-memory wasm bytes
        let wasm_bytes = wat2wasm(
            br#"(module
            (export "fib" (func $fib))
            (func $fib (param $n i32) (result i32)
             (if
              (i32.lt_s
               (local.get $n)
               (i32.const 2)
              )
              (then
               (return (i32.const 1))
              )
             )
             (return
              (i32.add
               (call $fib
                (i32.sub
                 (local.get $n)
                 (i32.const 2)
                )
               )
               (call $fib
                (i32.sub
                 (local.get $n)
                 (i32.const 1)
                )
               )
              )
             )
            )
           )
        "#,
        )?;
        let mut vm = vm.register_module_from_bytes("extern", wasm_bytes)?;

        // run `fib` function in the named module instance
        let returns = vm.run_func(Some("extern"), "fib", params!(10))?;
        assert_eq!(returns.len(), 1);
        assert_eq!(returns[0].to_i32(), 89);
    }

    Ok(())
}

Implementations

impl<'inst, T: ?Sized + Send + AsyncInst> Vm<'inst, T>

pub fn new(store: Store<'inst, T>) -> Self

pub fn register_module( &mut self, mod_name: Option<&str>, module: Module, ) -> WasmEdgeResult<&mut Self>

Registers a wasm module into this vm as a named or active module instance.

Arguments

mod_name - The exported name for the registered module. If None, then the module is registered as an active instance.

module - The module to be registered.

Error

If fail to register the given module, then an error is returned.

pub async fn run_func( &mut self, mod_name: Option<&str>, func_name: impl AsRef<str>, args: impl IntoIterator<Item = WasmValue> + Send, ) -> WasmEdgeResult<Vec<WasmValue>>

Runs an exported wasm function in a (named or active) module instance.

Arguments

mod_name - The exported name of the module instance, which holds the target function. If None, then the active module is used.

func_name - The exported name of the target wasm function.

args - The arguments to be passed to the target wasm function.

Error

If fail to run the wasm function, then an error is returned.

pub async fn run_func_with_timeout( &mut self, mod_name: Option<&str>, func_name: impl AsRef<str>, args: impl IntoIterator<Item = WasmValue> + Send, timeout: Duration, ) -> WasmEdgeResult<Vec<WasmValue>>

Runs an exported wasm function in a (named or active) module instance with a timeout setting

Arguments

mod_name - The exported name of the module instance, which holds the target function. If None, then the active module is used.

func_name - The exported name of the target wasm function.

args - The arguments to be passed to the target wasm function.

timeout - The maximum execution time of the function to be run.

Error

If fail to run the wasm function, then an error is returned.

pub fn store(&self) -> &Store<'inst, T>

Returns a reference to the internal store from this vm.

pub fn store_mut(&mut self) -> &mut Store<'inst, T>

Returns a mutable reference to the internal store from this vm.

pub fn active_module(&self) -> Option<&Instance>

Returns a reference to the active module instance from this vm.

Error

If fail to get the reference to the active module instance, then an error is returned.

pub fn active_module_mut(&mut self) -> Option<&mut Instance>

Returns a mutable reference to the active module instance from this vm.

Error

If fail to get the mutable reference to the active module instance, then an error is returned.

pub fn contains_module(&self, mod_name: impl AsRef<str>) -> bool

Checks if the vm contains a named module instance.

Argument

mod_name - The exported name of the target module instance.

pub fn named_instance_count(&self) -> usize

Returns the count of the named module instances this vm holds.

pub fn instance_names(&self) -> Vec<String>

Returns the names of all named module instances this vm holds.