Struct wasmedge_sdk::config::StatisticsConfigOptions

pub struct StatisticsConfigOptions { /* private fields */ }

Defines a group of the statistics configuration options.

StatisticsConfigOptions is used to set the statistics configuration options, which are

count_instructions determines if measuring the count of instructions when running a compiled or pure WASM.

measure_cost determines if measuring the instruction costs when running a compiled or pure WASM.

measure_time determines if measuring the running time when running a compiled or pure WASM.

Implementations

impl StatisticsConfigOptions

pub fn new() -> Self

Creates a new instance of StatisticsConfigOptions.

pub fn count_instructions(self, enable: bool) -> Self

Sets the instruction counting option.

Argument

enable specifies if support instruction counting or not when execution after AOT compilation.

pub fn measure_cost(self, enable: bool) -> Self

Sets the cost measuring option.

Argument

enable specifies if support cost measuring or not when execution after AOT compilation.

pub fn measure_time(self, enable: bool) -> Self

Sets the time measuring option.

Argument

enable specifies if support time measuring or not when execution after AOT compilation.