Struct wasmedge_sdk::dock::VmDock

pub struct VmDock { /* private fields */ }

Extends a Vm instance by supporting function arguments of Rust built-in types.

Implementations

impl VmDock

pub fn new(vm: Vm) -> Self

Creates a new VmDock to be associated with the given Vm.

Arguments

vm - The Vm instance to be extended.

pub fn run_func( &self, func_name: impl AsRef<str>, params: Vec<Param<'_>> ) -> WasmEdgeResult<Result<Vec<Box<dyn Any + Send + Sync>>, String>>

Runs an exported WASM function registered in a named or active module.

Arguments

func_name - The name of the exported WASM function to run.

params - The parameter values passed to the exported WASM function.

Error

If fail to run the WASM function, then an error is returned.