Enum wasmedge_sdk::ValType

pub enum ValType {
    I32,
    I64,
    F32,
    F64,
    V128,
    FuncRef,
    ExternRef,
}

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge value types.

Variants

I32

32-bit integer.

Integers are not inherently signed or unsigned, their interpretation is determined by individual operations.

I64

64-bit integer.

Integers are not inherently signed or unsigned, their interpretation is determined by individual operations.

F32

32-bit floating-point data as defined by the IEEE 754-2019.

F64

64-bit floating-point data as defined by the IEEE 754-2019.

V128

128-bit vector of packed integer or floating-point data.

The packed data can be interpreted as signed or unsigned integers, single or double precision floating-point values, or a single 128 bit type. The interpretation is determined by individual operations.

FuncRef

A reference to a host function.

ExternRef

A reference to object.