Module wasmedge_sdk::plugin

Defines plugin related structs.

Modules

ffi

Defines low-level types used in Plugin development.

Structs

NNPreloadwasi_nn

Preload config for initializing the wasi_nn plug-in.

Plugin

Represents a loaded plugin. It provides the APIs for accessing the plugin.

PluginDescriptor

Represents Plugin descriptor for plugins.

PluginManager

Defines the API to manage plugins.

PluginModule

Defines an import object that contains the required import data used when instantiating a module.

PluginModuleBuilder

Creates a plugin module.

PluginVersion

Defines the version of a plugin.

Enums

ExecutionTargetwasi_nn

Define where the graph should be executed.

GraphEncodingwasi_nn

Describes the encoding of the graph.

Type Aliases

ModuleInstanceCreateFn

Defines the type of the function that creates a module instance for a plugin.

PluginInstance

ProgramOptionType

Defines the type of the program options.