wasmedge_sdk
Struct Store

pub struct Store<'inst, T: ?Sized> { /* private fields */ }

The Store is a collection of registered modules and assists wasm modules in finding the import modules they need.

Implementations

impl<'inst, T: AsInstance + ?Sized> Store<'inst, T>

pub fn new( config: Option<&Config>, instances: HashMap<String, &'inst mut T>, ) -> WasmEdgeResult<Self>

Creates a new Store.

Error

If fail to create a new Store, then an error is returned.

pub fn register_active_module( &mut self, module: &Module, ) -> WasmEdgeResult<Instance>

Registers and instantiates a WasmEdge compiled module into this store as an anonymous active module instance, and returns the module instance.

Arguments

executor - The executor that runs the host functions in this store.

module - The validated module to be registered.

Error

If fail to register the given module, then an error is returned.

pub fn register_named_module( &mut self, name: impl AsRef<str>, module: &Module, ) -> WasmEdgeResult<()>

pub fn named_instance_count(&self) -> usize

Returns the number of the named module instances in this store.

pub fn instance_names(&self) -> Vec<String>

Returns the names of all registered named module instances.

pub fn contains(&self, mod_name: impl AsRef<str>) -> bool

Checks if the store contains a named module instance.

Argument

mod_name - The name of the named module.

pub fn get_instance_and_executor( &mut self, mod_name: impl AsRef<str>, ) -> Option<(&mut T, &mut Executor)>

pub fn get_named_wasm_and_executor( &mut self, mod_name: impl AsRef<str>, ) -> Option<(&mut Instance, &mut Executor)>

pub fn executor(&mut self) -> &mut Executor