Struct wasmedge_sdk::TableType

pub struct TableType { /* private fields */ }

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge TableType.

A TableType is used to declare the element type and the size range of a WasmEdge Table to be created.

Implementations

impl TableType

pub fn new(elem_ty: RefType, min: u32, max: Option<u32>) -> TableType

Creates a new TableType with the given element type and the size range.

Arguments

elem_ty - The element type of the table to be created.

min - The minimum size of the table to be created.

max - The maximum size of the table to be created.

pub fn elem_ty(&self) -> RefType

Returns the element type defined in the TableType.

pub fn minimum(&self) -> u32

Returns the minimum size defined in the TableType.

pub fn maximum(&self) -> Option<u32>

Returns the maximum size defined in the TableType.