wasmedge_sdk
Trait WasmValTypeList

pub trait WasmValTypeList
where
    Self: Sized,
{
    type Array: AsMut<[i128]>;

    // Required method
    fn wasm_types() -> &'static [ValType];
}

Describes the mapping of a tuple of Rust types to Wasm types.

use wasmedge_sdk::{types::ExternRef, ValType, WasmValTypeList};

assert_eq!(
     <(i32, i64, f32, f64, ExternRef)>::wasm_types(),
     [ValType::I32, ValType::I64, ValType::F32, ValType::F64, ValType::ExternRef]
);

Required Associated Types

type Array: AsMut<[i128]>

The array type that can hold all the represented values.

Note that all values are stored in their binary form.

Required Methods

fn wasm_types() -> &'static [ValType]

Get the Wasm types for the tuple (list) of currently represented values.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.