Struct wasmedge_sdk::wasi::WasiInstance

pub struct WasiInstance { /* private fields */ }

Represents a wasi module instance.

Implementations

impl WasiInstance

pub fn initialize(
&mut self,
args: Option<Vec<&str>>,
envs: Option<Vec<&str>>,
preopens: Option<Vec<&str>>
)

Initializes the WASI host module with the given parameters.

Arguments

args - The commandline arguments. The first argument is the program name.

envs - The environment variables in the format ENV_VAR_NAME=VALUE.

preopens - The directories to pre-open. The required format is DIR1:DIR2.

pub fn exit_code(&self) -> u32

Returns the WASI exit code.

The WASI exit code can be accessed after running the “_start” function of a wasm32-wasi program.

pub fn native_handler(&self, fd: i32) -> WasmEdgeResult<u64>

Returns the native handler from the mapped FD/Handler.

Argument

fd - The WASI mapped Fd.

Error

If fail to get the native handler, then an error is returned.