Crate wasmedge_sdk

Overview

The wasmedge-sdk crate defines a group of high-level Rust APIs, which are used to build up business applications.

Notice that wasmedge-sdk requires Rust v1.63 or above in the stable channel.

Versioning Table

The following table provides the versioning information about each crate of WasmEdge Rust bindings.

wasmedge-sdk	WasmEdge lib	wasmedge-sys	wasmedge-types	wasmedge-macro

0.7.1	0.11.2	0.12.2	0.3.1	0.3.0

0.7.0	0.11.2	0.12	0.3.1	0.3.0

0.6.0	0.11.2	0.11	0.3.0	0.2.0

0.5.0	0.11.1	0.10	0.3.0	0.1.0

0.4.0	0.11.0	0.9	0.2.1	-

0.3.0	0.10.1	0.8	0.2	-

0.1.0	0.10.0	0.7	0.1	-

Build

To use or build the wasmedge-sdk crate, the WasmEdge library is required.

If you choose to use install.sh to install WasmEdge Runtime on your local system. Please use WASMEDGE_INCLUDE_DIR and WASMEDGE_LIB_DIR to specify the paths to the include and lib directories, respectively. For example, use the following commands to specify the paths after using bash install.sh --path=$HOME/wasmedge-install to install WasmEdge Runtime on Ubuntu 20.04:

export WASMEDGE_INCLUDE_DIR=$HOME/wasmedge-install/include
export WASMEDGE_LIB_DIR=$HOME/wasmedge-install/lib

If you choose to manually download WasmEdge Runtime binary from WasmEdge Releases Page, it is strongly recommended to place it in $HOME/.wasmedge directory. It looks like below on Ubuntu 20.04. wasmedge-sdk will search the directory automatically, you do not have to set any environment variables for it.

// $HOME/.wasmedge/
.
|-- bin
|   |-- wasmedge
|   `-- wasmedgec
|-- include
|   `-- wasmedge
|       |-- enum.inc
|       |-- enum_configure.h
|       |-- enum_errcode.h
|       |-- enum_types.h
|       |-- int128.h
|       |-- version.h
|       `-- wasmedge.h
`-- lib64
    |-- libwasmedge_c.so
    `-- wasmedge
        `-- libwasmedgePluginWasmEdgeProcess.so

5 directories, 11 files

Enable WasmEdge Plugins

If you’d like to enable WasmEdge Plugins (currently, only available on Linux platform), please use WASMEDGE_PLUGIN_PATH environment variable to specify the path to the directory containing the plugins. For example, use the following commands to specify the path on Ubuntu 20.04:

export WASMEDGE_PLUGIN_PATH=$HOME/.wasmedge/lib/wasmedge

Example

The example below is using wasmedge-sdk to run a WebAssembly module written with its WAT format (textual format):

// If the version of rust used is less than v1.63, please uncomment the follow attribute.
// #![feature(explicit_generic_args_with_impl_trait)]

use wasmedge_sdk::{Executor, FuncTypeBuilder, ImportObjectBuilder, Module, Store, error::HostFuncError, WasmValue, wat2wasm, Caller, host_function};
 
#[cfg_attr(test, test)]
fn main() -> anyhow::Result<()> {
    let wasm_bytes = wat2wasm(
        br#"
(module
  ;; First we define a type with no parameters and no results.
  (type $no_args_no_rets_t (func (param) (result)))
 
  ;; Then we declare that we want to import a function named "env" "say_hello" with
  ;; that type signature.
  (import "env" "say_hello" (func $say_hello (type $no_args_no_rets_t)))
 
  ;; Finally we create an entrypoint that calls our imported function.
  (func $run (type $no_args_no_rets_t)
    (call $say_hello))
  ;; And mark it as an exported function named "run".
  (export "run" (func $run)))
"#,
    )?;
 
    // We define a function to act as our "env" "say_hello" function imported in the
    // Wasm program above.
    #[host_function]
    fn say_hello_world(_: Caller, _: Vec<WasmValue>) -> Result<Vec<WasmValue>, HostFuncError> {
        println!("Hello, world!");
 
        Ok(vec![])
    }
 
    // create an import module
    let import = ImportObjectBuilder::new()
        .with_func::<(), ()>("say_hello", Box::new(say_hello_world))?
        .build("env")?;
 
    // loads a wasm module from the given in-memory bytes
    let module = Module::from_bytes(None, &wasm_bytes)?;
 
    // create an executor
    let mut executor = Executor::new(None, None)?;
 
    // create a store
    let mut store = Store::new()?;
 
    // register the module into the store
    store.register_import_module(&mut executor, &import)?;
    let extern_instance = store.register_named_module(&mut executor, "extern", &module)?;
 
    // get the exported function "run"
    let run = extern_instance.func("run").ok_or(anyhow::Error::msg(
        "Not found exported function named 'run'.",
    ))?;
 
    // run host function
    run.call(&mut executor, [])?;
 
    Ok(())
}

[Click for more examples]

See also

WasmEdge Runtime

WasmEdge C API Documentation

Modules

config

Defines the structs used to construct configurations.

dock

Defines WasmEdge VmDock and Param types.

error

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge error types.

types

Defines the general types.

wasi

Defines wasi module instance types, including WasiInstance, WasiNnInstance, wasi-crypto instances.

Macros

params

Used to pass arguments to Func::call, FuncRef::call, Engine::run_func, or Engine::run_func_ref. Notice that to use the macro, it is required to use WasmVal trait. If the version of rust used is less than v1.63, please place #![feature(explicit_generic_args_with_impl_trait)] on the root of the program.

Structs

Caller

Represents the calling frame on top of stack.

CoreVersion

The version info of WasmEdge core

Driver

Defines WasmEdge Driver functions

Executor

Defines an execution environment for both pure WASM and compiled WASM.

ExportType

Defines the types of the exported instances.

Func

Defines a host function instance.

FuncRef

Defines a reference to a host function.

FuncType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge FuncType.

FuncTypeBuilder

Defines a type builder for creating a FuncType instance.

Global

Defines a WebAssembly global variable, which stores a single value of the given GlobalType and a flag indicating whether it is mutable or not.

GlobalType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge GlobalType.

ImportObject

Defines an import object that contains the required import data used when instantiating a module.

ImportObjectBuilder

Creates a normal, wasi, or wasmedge process import object.

ImportType

Defines the types of the imported instances.

Instance

Represents an instantiated module.

LogManager

Manipulates the runtime logger.

Memory

Defines a linear memory.

MemoryType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge MemoryType.

Module

Defines compiled in-memory representation of an input WASM binary.

PluginManager

Defines the API to manage plugins.

Statistics

Used to collect statistics of the WasmEdge runtime, such as the count of instructions in execution.

Store

Represents all global state that can be manipulated by WebAssembly programs. A store consists of the runtime representation of all instances of functions, tables, memories, and globals.

Table

Defines a table storing the references to host functions or external objects.

TableType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Struct of WasmEdge TableType.

Vm

A Vm defines a virtual environment for managing WebAssembly programs.

WasmEdgeProcessInstance

Represents a wasmedge_process module instance.

Enums

CompilerOptimizationLevel

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge AOT compiler optimization level.

CompilerOutputFormat

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge AOT compiler output binary format.

ExternalInstanceType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines the type of external WasmEdge instances.

Mutability

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines the mutability property of WasmEdge Global variables.

RefType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge reference types.

ValType

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Defines WasmEdge value types.

Traits

AsInstance

The object used as an module instance is required to implement this trait.

Engine

The object that is used to perform a host function is required to implement this trait.

WasmVal

Defines the function converting a value of Rust type to the one of Wasm type.

WasmValType

Describes the mapping of Rust type to Wasm type.

WasmValTypeList

Describes the mapping of a tuple of Rust types to Wasm types.

Functions

wat2wasm

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module.

Type Definitions

CallingFrame

WasmEdgeResult

Parses in-memory bytes as either the WebAssembly Text format, or a binary WebAssembly module The WasmEdge result type.

WasmValue

WebAssembly value type.

Attribute Macros

async_host_function

Declare a native async function that will be used to create an async host function instance.

host_function

Declare a native function that will be used to create a host function instance.