Crate wasmedge_sdk

Overview

The wasmedge-sdk crate defines a group of high-level Rust APIs, which are used to build up business applications.

Notice that wasmedge-sdk requires the nightly version of Rust.

Usage

To use or build the wasmedge-sdk crate, the wasmedge-core is required. The required header files, library and plugins should be placed in $HOME/.wasmedge/ directory. The directory structure on macOS looks like below:

// $HOME/.wasmedge/
.
├── include
│   └── wasmedge
│       ├── dense_enum_map.h
│       ├── enum.inc
│       ├── enum_configure.h
│       ├── enum_errcode.h
│       ├── enum_types.h
│       ├── int128.h
│       ├── spare_enum_map.h
│       ├── version.h
│       └── wasmedge.h
├── lib
│   └── libwasmedge_c.dylib
└── plugin
    └── libwasmedgePluginWasmEdgeProcess.dylib

A quick-start example

The example below is using wasmedge-sdk to run a WebAssembly module written with its WAT format (textual format):

#![feature(explicit_generic_args_with_impl_trait)]

use wasmedge_sdk::{Executor, FuncTypeBuilder, ImportObjectBuilder, Module, Store};
use wasmedge_sys::WasmValue;
use wasmedge_types::wat2wasm;
 
#[cfg_attr(test, test)]
fn main() -> anyhow::Result<()> {
    let wasm_bytes = wat2wasm(
        br#"
(module
  ;; First we define a type with no parameters and no results.
  (type $no_args_no_rets_t (func (param) (result)))
 
  ;; Then we declare that we want to import a function named "env" "say_hello" with
  ;; that type signature.
  (import "env" "say_hello" (func $say_hello (type $no_args_no_rets_t)))
 
  ;; Finally we create an entrypoint that calls our imported function.
  (func $run (type $no_args_no_rets_t)
    (call $say_hello))
  ;; And mark it as an exported function named "run".
  (export "run" (func $run)))
"#,
    )?;
 
    // We define a function to act as our "env" "say_hello" function imported in the
    // Wasm program above.
    fn say_hello_world(_inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u8> {
        println!("Hello, world!");
 
        Ok(vec![])
    }
 
    // create an import module
    let import = ImportObjectBuilder::new()
        .with_func::<(), ()>("say_hello", Box::new(say_hello_world))?
        .build("env")?;
 
    // loads a wasm module from the given in-memory bytes
    let module = Module::from_bytes(None, &wasm_bytes)?;
 
    // create an executor
    let mut executor = Executor::new(None, None)?;
 
    // create a store
    let mut store = Store::new()?;
 
    // register the module into the store
    store.register_import_module(&mut executor, &import)?;
    let extern_instance = store.register_named_module(&mut executor, "extern", &module)?;
 
    // get the exported function "run"
    let run = extern_instance.func("run").ok_or(anyhow::Error::msg(
        "Not found exported function named 'run'.",
    ))?;
 
    // run host function
    run.call(&mut executor, [])?;
 
    Ok(())
}

[Click for more examples]

See also

WasmEdge Runtime

WasmEdge C API Documentation

wasmedge-sys: WasmEdge Low-level Rust APIs

wasmedge-types: WasmEdge Types

Modules

config

Defines the structs used to construct configurations.

types

Defines the general types.

Macros

params

Used to pass arguments to Func::call, FuncRef::call, Engine::run_func, or Engine::run_func_ref. Notice that to use the macro, it is required to use WasmVal trait and place #![feature(explicit_generic_args_with_impl_trait)] on the root of the program.

Structs

Compiler

Defines WasmEdge ahead-of-time(AOT) compiler and the relevant APIs.

Executor

Defines an execution environment for both pure WASM and compiled WASM.

ExportType

Defines the types of the exported instances.

Func

Defines a host function instance.

FuncRef

Defines a reference to a host function.

FuncTypeBuilder

Defines a type builder for creating a FuncType instance.

Global

Defines a WebAssembly global variable, which stores a single value of the given GlobalType and a flag indicating whether it is mutable or not.

ImportObject

Defines an import object that contains the required import data used when instantiating a module.

ImportObjectBuilder

Creates a normal, wasi, or wasmedge process import object.

ImportType

Defines the types of the imported instances.

Instance

Represents an instantiated module.

Memory

Defines a linear memory.

Module

Defines compiled in-memory representation of an input WASM binary.

Statistics

Used to collect statistics of the WasmEdge runtime, such as the count of instructions in execution.

Store

Represents all global state that can be manipulated by WebAssembly programs. A store consists of the runtime representation of all instances of functions, tables, memories, and globals.

Table

Defines a table storing the references to host functions or external objects.

Vm

A Vm defines a virtual environment for managing WebAssembly programs.

WasiInstance

Represents a wasi module instance.

WasmEdgeProcessInstance

Represents a wasmedge_process module instance.

Traits

Engine

The object that is used to perform a host function is required to implement this trait.

WasmVal

Defines the function converting a value of Rust type to the one of Wasm type.

WasmValType

Describes the mapping of Rust type to Wasm type.

WasmValTypeList

Describes the mapping of a tuple of Rust types to Wasm types.