Trait wasmedge_sdk::WasmValTypeList

pub trait WasmValTypeListwhere
    Self: Sized,{
    type Array: AsMut<[i128]>;

    fn wasm_types() -> &'static [ValType];
}

Describes the mapping of a tuple of Rust types to Wasm types.

use wasmedge_sdk::{FuncRef, types::ExternRef, ValType, WasmValTypeList};

assert_eq!(
     <(i32, i64, f32, f64, FuncRef, ExternRef)>::wasm_types(),
     [ValType::I32, ValType::I64, ValType::F32, ValType::F64, ValType::FuncRef, ValType::ExternRef]
);

Required Associated Types

type Array: AsMut<[i128]>

The array type that can hold all the represented values.

Note that all values are stored in their binary form.

Required Methods

fn wasm_types() -> &'static [ValType]

Get the Wasm types for the tuple (list) of currently represented values.